Simplify mongo instrumentation and fix checkpoint emitting
