+++ dd-java-agent/instrumentation/mongo/common/common.gradle
@@ -1,0 +1,6 @@
+apply from: "$rootDir/gradle/java.gradle"
+
+dependencies {
+  compileOnly group: 'org.mongodb', name: 'mongo-java-driver', version: '3.1.0'
+  testImplementation project(':dd-java-agent:instrumentation:mongo').sourceSets.test.output
+}
