+++ dd-java-agent/instrumentation/mongo/common/src/main/java/datadog/trace/instrumentation/mongo/BsonScrubber.java
@@ -1,0 +1,12 @@
+package datadog.trace.instrumentation.mongo;
+
+import org.bson.BsonReader;
+
+/** A shared interface for the bson scrubbers used in the mongo instrumentations */
+public interface BsonScrubber extends AutoCloseable {
+  void pipe(BsonReader reader);
+
+  void close();
+
+  String getResourceName();
+}
