+++ dd-java-agent/instrumentation/mongo/driver-3.1/driver-3.1.gradle
@@ -5,6 +5,27 @@ muzzle {
     versions = "[3.1,)"
     assertInverse = true
   }
+  pass {
+    group = "org.mongodb"
+    module = "mongodb-driver-async"
+    versions = "[3.0,)"
+    extraDependency 'org.mongodb:mongo-java-driver:3.1.0'
+    assertInverse = true
+  }
+  pass {
+    group = "org.mongodb"
+    module = "mongodb-driver-sync"
+    versions = "[3.0,)"
+    extraDependency 'org.mongodb:mongo-java-driver:3.1.0'
+    assertInverse = true
+  }
+  pass {
+    group = "org.mongodb"
+    module = "mongodb-driver-core"
+    versions = "[3.1,)"
+    extraDependency 'org.mongodb:mongo-java-driver:3.1.0'
+    assertInverse = true
+  }
 }
 
 apply from: "$rootDir/gradle/java.gradle"
@@ -19,10 +40,18 @@ testSets {
 
 dependencies {
   compileOnly group: 'org.mongodb', name: 'mongo-java-driver', version: '3.1.0'
+  compileOnly group: 'org.mongodb', name: 'mongodb-driver-core', version: '3.3.0'
+  compileOnly group: 'org.mongodb', name: 'mongodb-driver-async', version: '3.3.0'
+
+  implementation(project(':dd-java-agent:instrumentation:mongo:common')) {
+    transitive = false
+  }
 
   testImplementation project(':dd-java-agent:instrumentation:mongo').sourceSets.test.output
   testImplementation group: 'de.flapdoodle.embed', name: 'de.flapdoodle.embed.mongo', version: '1.50.5'
 
   testImplementation group: 'org.mongodb', name: 'mongo-java-driver', version: '3.1.0'
   latestDepTestImplementation group: 'org.mongodb', name: 'mongo-java-driver', version: '+'
+  testImplementation group: 'org.mongodb', name: 'mongodb-driver-async', version: '3.3.0'
+  latestDepTestImplementation group: 'org.mongodb', name: 'mongodb-driver-async', version: '+'
 }
