--- dd-java-agent/instrumentation/mongo/driver-3.1/src/main/java/datadog/trace/instrumentation/mongo/MongoClientInstrumentation.java
@@ -1,73 +1,0 @@
-package datadog.trace.instrumentation.mongo;
-
-import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
-import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.declaresField;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.isPublic;
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
-
-import com.google.auto.service.AutoService;
-import com.mongodb.event.CommandListener;
-import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.bootstrap.InstrumentationContext;
-import java.util.List;
-import java.util.Map;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-import org.bson.BsonDocument;
-import org.bson.ByteBuf;
-
-@AutoService(Instrumenter.class)
-public final class MongoClientInstrumentation extends Instrumenter.Tracing {
-
-  public MongoClientInstrumentation() {
-    super("mongo");
-  }
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("com.mongodb.MongoClientOptions$Builder")
-        .and(declaresField(named("commandListeners")));
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      packageName + ".MongoClientDecorator",
-      packageName + ".BsonScrubber",
-      packageName + ".BsonScrubber$1",
-      packageName + ".BsonScrubber$2",
-      packageName + ".Context",
-      packageName + ".TracingCommandListener"
-    };
-  }
-
-  @Override
-  public Map<String, String> contextStore() {
-    return singletonMap("org.bson.BsonDocument", "org.bson.ByteBuf");
-  }
-
-  @Override
-  public void adviceTransformations(AdviceTransformation transformation) {
-    transformation.applyAdvice(
-        isMethod().and(isPublic()).and(named("build")).and(takesArguments(0)),
-        MongoClientInstrumentation.class.getName() + "$MongoClientAdvice");
-  }
-
-  public static class MongoClientAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void injectTraceListener(
-        @Advice.FieldValue("commandListeners") List<CommandListener> listeners) {
-      if (!listeners.isEmpty()
-          && listeners.get(listeners.size() - 1).getClass().getName().startsWith("datadog.")) {
-        return;
-      }
-      listeners.add(
-          new TracingCommandListener(
-              InstrumentationContext.get(BsonDocument.class, ByteBuf.class)));
-    }
-  }
-}
