+++ dd-java-agent/instrumentation/mongo/driver-3.1/src/main/java/datadog/trace/instrumentation/mongo/MongoDecorator31.java
@@ -1,0 +1,10 @@
+package datadog.trace.instrumentation.mongo;
+
+public final class MongoDecorator31 extends MongoDecorator {
+  public static final MongoDecorator31 INSTANCE = new MongoDecorator31();
+
+  @Override
+  protected BsonScrubber newScrubber() {
+    return new BsonScrubber31();
+  }
+}
