+++ dd-java-agent/instrumentation/mongo/driver-3.4/driver-3.4.gradle
@@ -1,0 +1,86 @@
+muzzle {
+  pass {
+    group = "org.mongodb"
+    module = "mongodb-driver-core"
+    versions = "[3.1,)"
+    assertInverse = true
+    extraDependency 'org.mongodb:mongo-java-driver:3.4.0'
+  }
+  fail {
+    group = "org.mongodb"
+    module = "mongodb-driver-core"
+    versions = "[3.1,)"
+    extraDependency 'org.mongodb:mongo-java-driver:3.3.0'
+  }
+  pass {
+    group = "org.mongodb"
+    module = "mongodb-driver-async"
+    versions = "[3.0,)"
+    extraDependency 'org.mongodb:mongo-java-driver:3.4.0'
+    assertInverse = true
+  }
+  fail {
+    group = "org.mongodb"
+    module = "mongodb-driver-async"
+    versions = "[3.0,)"
+    extraDependency 'org.mongodb:mongo-java-driver:3.1.0'
+  }
+  pass {
+    group = "org.mongodb"
+    module = "mongodb-driver-sync"
+    versions = "[3.0,)"
+    extraDependency 'org.mongodb:mongo-java-driver:3.4.0'
+    assertInverse = true
+  }
+  fail {
+    group = "org.mongodb"
+    module = "mongodb-driver-sync"
+    versions = "[3.0,)"
+    extraDependency 'org.mongodb:mongo-java-driver:3.1.0'
+  }
+  pass {
+    group = "org.mongodb"
+    module = "mongo-java-driver"
+    versions = "[3.4,)"
+  }
+  fail {
+    group = "org.mongodb"
+    module = "mongo-java-driver"
+    versions = "[,3.4)"
+  }
+}
+
+apply from: "$rootDir/gradle/java.gradle"
+
+apply plugin: 'org.unbroken-dome.test-sets'
+
+testSets {
+  latestDepTest {
+    dirName = 'test'
+  }
+}
+
+dependencies {
+  compileOnly group: 'org.mongodb', name: 'mongo-java-driver', version: '3.4.0'
+  compileOnly group: 'org.mongodb', name: 'mongodb-driver-core', version: '3.4.0'
+  compileOnly group: 'org.mongodb', name: 'mongodb-driver-async', version: '3.4.0'
+
+  implementation(project(':dd-java-agent:instrumentation:mongo:common')) {
+    transitive = false
+  }
+
+  testImplementation project(':dd-java-agent:instrumentation:mongo').sourceSets.test.output
+  testImplementation group: 'de.flapdoodle.embed', name: 'de.flapdoodle.embed.mongo', version: '1.50.5'
+
+  // We need to pull in this dependency to get the 'suspend span' instrumentation for spock tests
+  // as well as to test the instrumentaiton 'layering' (3.4 instrumentation should take precedence
+  // over 3.1 - otherwise the tests here should fail)
+  testImplementation(project(':dd-java-agent:instrumentation:mongo:driver-3.1')) {
+    transitive = false
+  }
+
+  testImplementation group: 'org.mongodb', name: 'mongo-java-driver', version: '3.4.0'
+  latestDepTestImplementation group: 'org.mongodb', name: 'mongo-java-driver', version: '3.+'
+  testImplementation group: 'org.mongodb', name: 'mongodb-driver-async', version: '3.4.0'
+  latestDepTestImplementation group: 'org.mongodb', name: 'mongodb-driver-async', version: '3.+'
+}
