+++ dd-java-agent/instrumentation/mongo/driver-3.4/src/main/java/datadog/trace/instrumentation/mongo/MongoDecorator34.java
@@ -1,0 +1,10 @@
+package datadog.trace.instrumentation.mongo;
+
+public final class MongoDecorator34 extends MongoDecorator {
+  public static final MongoDecorator34 INSTANCE = new MongoDecorator34();
+
+  @Override
+  protected BsonScrubber newScrubber() {
+    return new BsonScrubber34();
+  }
+}
