+++ dd-java-agent/instrumentation/mongo/driver-4-test/driver-4-test.gradle
@@ -1,0 +1,37 @@
+// Set properties before any plugins get loaded
+ext {
+  // Mongo4 driver requires Java 8 minimum
+  minJavaVersionForTests = JavaVersion.VERSION_1_8
+}
+
+apply from: "$rootDir/gradle/java.gradle"
+
+apply plugin: 'org.unbroken-dome.test-sets'
+
+testSets {
+  latestDepTest {
+    dirName = 'test'
+  }
+}
+
+dependencies {
+  testImplementation(project(':dd-java-agent:instrumentation:mongo:common')) {
+    transitive = false
+  }
+  testImplementation(project(':dd-java-agent:instrumentation:mongo:driver-3.1')) {
+    transitive = false
+  }
+  testImplementation(project(':dd-java-agent:instrumentation:mongo:driver-3.4')) {
+    transitive = false
+  }
+
+  testImplementation project(':dd-java-agent:instrumentation:mongo').sourceSets.test.output
+  testImplementation group: 'de.flapdoodle.embed', name: 'de.flapdoodle.embed.mongo', version: '1.50.5'
+
+  testImplementation group: 'org.mongodb', name: 'bson', version: '4.0.0'
+  testImplementation group: 'org.mongodb', name: 'mongodb-driver-sync', version: '4.0.0'
+  testImplementation group: 'org.mongodb', name: 'mongodb-driver-reactivestreams', version: '4.0.0'
+  latestDepTestImplementation group: 'org.mongodb', name: 'bson', version: '4.+'
+  latestDepTestImplementation group: 'org.mongodb', name: 'mongodb-driver-sync', version: '4.2+'
+  latestDepTestImplementation group: 'org.mongodb', name: 'mongodb-driver-reactivestreams', version: '4.2+'
+}
