+++ settings.gradle
@@ -184,9 +184,10 @@ include ':dd-java-agent:instrumentation:log4j2'
 include ':dd-java-agent:instrumentation:logback-1'
 include ':dd-java-agent:instrumentation:micronaut-http-server-netty-2'
 include ':dd-java-agent:instrumentation:mongo'
+include ':dd-java-agent:instrumentation:mongo:common'
 include ':dd-java-agent:instrumentation:mongo:driver-3.1'
-include ':dd-java-agent:instrumentation:mongo:driver-4'
-include ':dd-java-agent:instrumentation:mongo:driver-async-3.3'
+include ':dd-java-agent:instrumentation:mongo:driver-3.4'
+include ':dd-java-agent:instrumentation:mongo:driver-4-test'
 include ':dd-java-agent:instrumentation:mule-4'
 include ':dd-java-agent:instrumentation:netty-3.8'
 include ':dd-java-agent:instrumentation:netty-4.0'
